+++ RELEASE_NOTES.md
@@ -1,4 +1,5 @@
 ### 0.11.1-beta
+* Updated internal `TypeShape` dependency to 10.0.0
 * Updated internal `AwaitTaskCorrect` implementation to align with [canonical version](http://www.fssnip.net/7Rc/title/AsyncAwaitTaskCorrect) [#49](https://github.com/fsprojects/FSharp.AWS.DynamoDB/pull/49) 
 * Added SourceLink info (using `DotNet.ReproducibleBuilds`)